HS-1302: Cleanup more detector related code
