--- metrics-processor/main/src/main/java/org/opennms/horizon/tsdata/detector/TaskSetDetectorResultProcessor.java
@@ -1,46 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2023 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2023 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.horizon.tsdata.detector;
-
-import org.opennms.taskset.contract.DetectorResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class TaskSetDetectorResultProcessor {
-
-    private static final Logger LOG = LoggerFactory.getLogger(TaskSetDetectorResultProcessor.class);
-
-    public void processDetectorResponse(String tenantId, String taskId, DetectorResponse detectorResponse) throws IOException {
-            LOG.info("Have detector response, tenant-id: {}; task-id={}; detected={}", tenantId, taskId, detectorResponse.getDetected());
-    }
-}
