+++ minion/minion-taskset/taskset-worker/src/main/java/org/opennms/horizon/minion/taskset/worker/TaskExecutionResultProcessor.java
@@ -2,7 +2,6 @@ package org.opennms.horizon.minion.taskset.worker;
 
 import org.opennms.horizon.minion.plugin.api.CollectionSet;
 import org.opennms.horizon.minion.plugin.api.ScanResultsResponse;
-import org.opennms.horizon.minion.plugin.api.ServiceDetectorResponse;
 import org.opennms.horizon.minion.plugin.api.ServiceMonitorResponse;
 
 public interface TaskExecutionResultProcessor {
@@ -14,13 +13,6 @@ public interface TaskExecutionResultProcessor {
      */
     void queueSendResult(String uuid, ScanResultsResponse scanResultsResponse);
 
-    /**
-     * Queue the given detector result to be sent out.
-     *
-     * @param uuid
-     * @param serviceDetectorResponse
-     */
-    void queueSendResult(String uuid, ServiceDetectorResponse serviceDetectorResponse);
 
     /**
      * Queue the given monitor result to be sent out.
