+++ minion/plugins/plugins-api/src/main/java/org/opennms/horizon/minion/plugin/api/ServiceDetector.java
@@ -7,7 +7,6 @@ import java.util.concurrent.CompletableFuture;
 
 public interface ServiceDetector {
 
-    CompletableFuture<ServiceDetectorResponse> detect(Any config, long nodeId);
 
     CompletableFuture<ServiceResult> detect(String host, Any config);
